# Stop manually SIMDing in `swap_nonoverlapping`