@scottmcm 2022-02-24
@bors retry network issue
```
warning: spurious network error (2 tries remaining): failed to get 200 response from `[https://crates.io/api/v1/crates/serde/1.0.125/download`,](https://crates.io/api/v1/crates/serde/1.0.125/download%60,) got 502
warning: spurious network error (2 tries remaining): failed to get 200 response from `[https://crates.io/api/v1/crates/regex/1.5.4/download`,](https://crates.io/api/v1/crates/regex/1.5.4/download%60,) got 502
warning: spurious network error (2 tries remaining): failed to get 200 response from `[https://crates.io/api/v1/crates/serde_json/1.0.59/download`,](https://crates.io/api/v1/crates/serde_json/1.0.59/download%60,) got 502
error: failed to download from `[https://crates.io/api/v1/crates/cfg-if/1.0.0/download`](https://crates.io/api/v1/crates/cfg-if/1.0.0/download%60)
```