@rust-log-analyzer 2022-02-24
The job **`aarch64-gnu`** failed! Check out the build log: [(web)](https://github.com/rust-lang-ci/rust/runs/5322096133?check_suite_focus=true) [(plain)](https://github.com/rust-lang-ci/rust/commit/c7e3ec112c881757bcd4d57840e009620c505b15/checks/5322096133/logs)

<details><summary><i>Click to see the possible cause of the failure (guessed by this bot)</i></summary>

```plain

```

</details>