@RalfJung 2022-02-25
Ah, I think this is an instance of https://github.com/rust-lang/rust/issues/69488 -- not a bug in this code, but a limitation in the Miri engine that is exposed by the new swap implementation.

We could add a `cfg(miri)` work-around, but the same issue also affects CTFE execution and that work-around won't help there.