@scottmcm 2022-02-25
Thanks for looking at this, Ralf.  I'm glad to hear that copying as `MaybeUninit<usize>` is theoretically ok, at least.

I suppose another option would be to skip this entirely with `const_eval_select`?  That'd probably be faster, since there's no autovectorization at MIR-level, not is there a worry about copying big things to `alloca`s in MIRI.

For a quick thing, I could make a PR to add `miri` to the `cfg` here, if you'd like?

https://github.com/rust-lang/rust/blob/6cbc6c35e4b0c948114619a1c883a75b731d32c5/library/core/src/mem/mod.rs#L711