@RalfJung 2022-02-25
> I'm glad to hear that copying as MaybeUninit<usize> is theoretically ok, at least.

That is the intention.
I can't vouch for LLVM though, sadly `poison` in LLVM is per-value, not per-byte, so a `MaybeUninit<usize>` (which becomes `i64`) loading a partially-poison value will be fully poison. This is, I think, a fundamental limitation of LLVM and one reason why it needs the "byte" type. (But sadly the LLVM community is far from convinced of this.)

> For a quick thing, I could make a PR to add miri to the cfg here, if you'd like?

How would that help? `swap_nonoverlapping` seems to still bottom out in `swap_simple`, which will still cause the same problem.

I *think* there is a fairly simple fix for https://github.com/rust-lang/rust/issues/69488, but I don't know its larger consequences. That and the LLVM concerns mentioned above made me hesitate. But it might be the right time to see if that fix works -- that would help both for CTFE and Miri.