@RalfJung 2022-02-25
I created a standalone testcase and opened an issue: https://github.com/rust-lang/rust/issues/94371